requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
mediapipe>=0.10.0
paho-mqtt>=1.6.0

